refactor(All): MVC 모델로 객체 구분
