+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.Controller;
+
 public class Application {
     public static void main(String[] args) {
         Controller controller = new Controller();
