+++ src/main/java/lotto/model/WinningLotto.java
@@ -1,7 +1,6 @@
-package lotto;
+package lotto.model;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
