+++ src/main/java/lotto/service/LottoResultCalculator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import lotto.dto.LottoResultDto;
