+++ src/main/java/lotto/service/RateOfReturnService.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
