+++ src/main/java/lotto/validation/Validation.java
@@ -1,10 +1,9 @@
-package lotto;
+package lotto.validation;
 
 public class Validation {
 
     public static void validateNumberValue(String inputLottoPurchaseMoney) {
         if (!inputLottoPurchaseMoney.matches("\\d+")) {
-//            System.out.println("[ERROR] 로또 구입 금액은 숫자를 입력해야 합니다.");
             throw new IllegalArgumentException("[ERROR] 로또 구입 금액은 숫자를 입력해야 합니다.");
         }
     }
