+++ src/test/java/lotto/RankTest.java
@@ -1,6 +1,9 @@
 package lotto;
 
 import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.Rank;
+import lotto.model.WinningLotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
