+++ src/test/java/lotto/WinningLottoTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.WinningLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
